Add serialize package and move io_utils module
This commit adds a new package called "serialize" and moves the "io_utils" module into it. This reorganization helps to better organize the codebase and improve code readability.
